# platform/api/openapi/v1/horecaos-api.storefront.json
{
  "components" : {
    "schemas" : {
...
      "CallbackRequestResponse" : {
        "properties" : {
          "leadId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "leadId" ],
        "type" : "object"
      },
...
      "CustomerCallbackRequest" : {
        "properties" : {
          "forCompany" : {
            "type" : "boolean"
          },
          "note" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          },
          "phone" : {
            "maxLength" : 32,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "phone" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Keycloak access token issued for the horecaos-api audience",
        "scheme" : "bearer",
        "type" : "http"
      },
      "storefrontAppId" : {
        "description" : "The registered app's id. An identifier, not a secret: it appears in a browser bundle by design. Sent alongside the customer's session, never instead of it.",
        "in" : "header",
        "name" : "X-Storefront-App-Id",
        "type" : "apiKey"
      },
      "storefrontAppSecret" : {
        "description" : "A confidential client's secret. Server-side only; a public client has none and must not send one.",
        "in" : "header",
        "name" : "X-Storefront-App-Secret",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
...
    "title" : "HorecaOS Platform API",
    "version" : "v1",
    "x-horecaos-contract" : {
      "appIdHeader" : "X-Storefront-App-Id",
      "appSecretHeader" : "X-Storefront-App-Secret",
      "deprecationWindowMonths" : 12,
      "publishedPathsNeverRemovedWithin" : "v1"
    }
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/storefront/tenants/{tenantId}/brands/{brandId}/me/callback-requests" : {
      "post" : {
        "description" : "Creates one lead in the call centre's queue, linked to the caller's own account. 202: what was created is a work item for a person, not a call. The number to ring is in the body. A guest asking on behalf of a company says so with forCompany=true, which files it as a catering enquiry.",
        "operationId" : "request",
        "parameters" : [ {
          "in" : "path",
          "name" : "tenantId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CustomerCallbackRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CallbackRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Ask the restaurant to call me back",
        "tags" : [ "Customer self-service" ],
        "x-horecaos-surface" : "published"
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ],
...
}